requirements.txt
amqp==5.3.1
asgiref==3.8.1
async-timeout==5.0.1
beautifulsoup4==4.13.4
billiard==4.2.1
celery==5.5.2
certifi==2025.4.26
charset-normalizer==3.4.1
click==8.1.8
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
Django==5.2
django-bootstrap5==25.1
django-excel-response2==3.0.6
django-jazzmin==3.0.1
django-six==1.0.5
django-unfold==0.59.0
djangorestframework==3.16.0
excel-base==1.0.4
idna==3.10
isoweek==1.3.3
kombu==5.5.3
pillow==11.2.1
prompt_toolkit==3.0.51
python-dateutil==2.9.0.post0
redis==5.2.1
requests==2.32.3
screen==1.0.1
six==1.17.0
soupsieve==2.7
sqlparse==0.5.3
TimeConvert==3.0.13
typing_extensions==4.13.2
tzdata==2025.2
tzlocal==5.3.1
urllib3==2.4.0
vine==5.1.0
wcwidth==0.2.13
xlwt==1.3.0